requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
asyncpg==0.27.0
redis==4.5.5
cryptography==41.0.1
python-dotenv==1.0.0